Make the project be aws lambda based (#108)
The service code is old and based on running things in heroku, lambda is
likely cheaper and easier for me to deal with, so I'm moving over to
that. I've learned a lot since I originally wrote that service code so
I'm going to start from scratch.

I've also made the java be more modular in case I end up wanting to
build more than one lambda for this project.