Fix the build after using LFS (#154)
It turns out you need to tell github actions that you are using LFS
otherwise it won't fully download the files and you end up pushing
pointers to images and not the images to  s3. This fixes that by telling
the UI build to actually download the artifacts for real.

Also I accidentially duplicated one of the articles, this removes the
extra.