Remove lambda (#160)
* Remove lambda

This repo has become the offthecob repo, and the work on the lambda
has been put on hold. The JVM part of the build is the slowest part by a
pretty big margin, I'm removing it. If I end up wanting to do this work
in the future, I'll likely have that code live in a different repo
anyway.

* remove more resources related to lambda

* remove additional cruft